Changed where the authentication code was to mainly reside in auth.py
